feat(starter/page): add custom error page
Reference: https://nextjs.org/docs/advanced-features/custom-error-page
