+++ packages/next-starter/src/app/components/index.ts
@@ -1,1 +1,2 @@
 export * from './MyApp';
+export * from './Error';
