+++ packages/next-starter/src/pages/_error.tsx
@@ -1,0 +1,3 @@
+import { Error } from '@app';
+
+export default Error;
